Move user and loginform to separate file
